+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `base_uri`/`base_url` property to sr.yaml. This can be a URI containing a base configuration for the sr.yaml, which is further extended by the contents of sr.yaml.
 - Fix `json_schema_url` not working with arbitrary URLs.
 - Add `json_schema_uri` alias for `json_schema_url`.
+- Rename `generator-file` embedded step to `generator`, making `generator-file` an alias.
 
 ## v0.8.0 (2022-11-07)